Remove NextJS boilerplate
